+++ js-web/vue-vite/apploader.js
@@ -1,3 +1,3 @@
-// copy this to projecrt root apploader.js
+// copy this to project root apploader.js
 import setup from './src/apps/web-template/setup.js'
 export default setup